requirements.txt
fastapi
uvicorn
supabase
dotenv
python-dotenv 
sqlalchemy 
bcrypt==4.0.1 
passlib
psycopg2-binary
python-multipart
pydantic
python-jose[cryptography]

requests
pytest
google-genai
